+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
-## [1.0.0] - 2022-01-27
+## [1.0.1] - 2023-02-17
+
+### Fixed
+
+- Fixed DIG handler, use default nameserver
+- Fixed PHPUnit tests data providers changed to static
+
+## [1.0.0] - 2023-01-27
 
 ### Added